moving permission requested alerts to SmartReceiptsActivity
