+++ app/src/main/java/co/smartreceipts/android/activities/SmartReceiptsActivity.java
@@ -1,13 +1,18 @@
 package co.smartreceipts.android.activities;
 
+import android.Manifest;
+import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
+import android.os.Build;
 import android.os.Bundle;
+import android.provider.Settings;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.widget.Toast;
 
 import androidx.annotation.NonNull;
+import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 
 import java.util.Set;
@@ -285,5 +290,33 @@ public class SmartReceiptsActivity extends AppCompatActivity implements HasAndro
     public void presentIntentImportFatalError() {
         Toast.makeText(this, R.string.toast_attachment_error, Toast.LENGTH_SHORT).show();
         finish();
+    }
+
+    private void alertAboutPermissions() {
+        final AlertDialog.Builder dlgAlert = new AlertDialog.Builder(this);
+        dlgAlert.setCancelable(false)
+                .setIcon(R.mipmap.ic_launcher)
+                .setTitle(getString(R.string.storage_permission_required))
+                .setMessage(getString(R.string.permission_must_be_granted))
+                .setNeutralButton(getString(R.string.ok), null).show();
+    }
+
+    private void alertChangePermissions(final Context context) {
+        final AlertDialog.Builder dlgAlert = new AlertDialog.Builder(context);
+        dlgAlert.setCancelable(false)
+                .setIcon(R.drawable.ic_error_outline_24dp)
+                .setTitle(getString(R.string.storage_permission_required))
+                .setMessage(getString(R.string.approve_permission))
+                .setNegativeButton(getString(R.string.no), null)
+                .setPositiveButton(getString(R.string.yes), (dialogInterface, i) -> {
+                    final Intent intent = new Intent();
+                    intent.setAction(Settings.ACTION_APPLICATION_DETAILS_SETTINGS);
+                    intent.addCategory(Intent.CATEGORY_DEFAULT);
+                    intent.setData(Uri.parse("package:" + context.getPackageName()));
+                    intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+                    intent.addFlags(Intent.FLAG_ACTIVITY_NO_HISTORY);
+                    intent.addFlags(Intent.FLAG_ACTIVITY_EXCLUDE_FROM_RECENTS);
+                    context.startActivity(intent);
+                }).show();
     }
 }
